% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{add.factors}
\alias{add.factors}
\title{Add factors to rki data}
\usage{
add.factors(dat)
}
\arguments{
\item{dat}{rki.de.data}
}
\value{
data frame with factors as columns for age.group, district, id.district, state, id.state
}
\description{
Add factors to rki data
}
