% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{age.plot.district}
\alias{age.plot.district}
\title{Plot age distribution of districts}
\usage{
age.plot.district(dat, var, filter.district, title)
}
\arguments{
\item{dat}{data}

\item{var}{variable to separate}

\item{filter.district}{keep only these}

\item{title}{title of plot}
}
\value{
ggplot
}
\description{
Plot age distribution of districts
}
