% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{age.plot.state}
\alias{age.plot.state}
\title{Plot age distribution of federal states}
\usage{
age.plot.state(dat, var, filter.state, title)
}
\arguments{
\item{dat}{data}

\item{var}{variable to separate}

\item{filter.state}{keep only these}

\item{title}{title of plot}
}
\value{
ggplot
}
\description{
Plot age distribution of federal states
}
