% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{de.nuts.mapping}
\alias{de.nuts.mapping}
\title{Local district id to nuts_3}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 412 rows and 3 columns.
}
\usage{
de.nuts.mapping
}
\description{
Local district id to nuts_3
}
\examples{
head(de.nuts.mapping)
}
\keyword{data}
