% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download.raw}
\alias{download.raw}
\title{Download call to ARCGIS}
\usage{
download.raw(
  offset = 0,
  max.record = 1000,
  exclude.ids = NULL,
  feature.server = "RKI_COVID19"
)
}
\arguments{
\item{offset}{offset on number of records for call}

\item{max.record}{maximum number of records to download in call}

\item{exclude.ids}{exclude ids from data}
}
\value{
data frame with data. Column names are translated from German
}
\description{
Download call to ARCGIS
}
\examples{
download.raw(feature.server = 'Covid19_RKI_Sums')
}
