% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download.state}
\alias{download.state}
\title{Download state specific data}
\usage{
download.state(existing.data = tibble::tibble(), max.record = 1000)
}
\arguments{
\item{existing.data}{data to be merged}

\item{max.record}{maximum number of records to download on each HTTP GET call}
}
\value{
data frame with all the data. Column name are translated from German
}
\description{
Download state specific data
}
