% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{identify.ranges}
\alias{identify.ranges}
\title{Algorithm to pick up a vector and find all ranges in it}
\usage{
\method{identify}{ranges}(dat, key.fun = range.write)
}
\arguments{
\item{dat}{vector of numbers}

\item{key.fun}{callback function to write ranges}
}
\value{
vector of strings defining all ranges in data
}
\description{
Algorithm to pick up a vector and find all ranges in it
}
\examples{
identify.ranges(c(1:3, 5, 7:8, 10:12, 14:15))
identify.ranges(c(1:3, 5, 7:8, 10:12, 14))
identify.ranges(c(1, 5, 7:8, 10:12, 14))
identify.ranges(c(1, 5, 7:8, 10:12, 14:19))
identify.ranges(rki.covid19 \%>\% pull(object.id) \%>\% unique \%>\% sort)
}
