% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{identify_ranges}
\alias{identify_ranges}
\title{Algorithm to pick up a vector and find all ranges in it}
\usage{
identify_ranges(dat, key.fun = range_write)
}
\arguments{
\item{dat}{vector of numbers}

\item{key.fun}{callback function to write ranges}
}
\value{
vector of strings defining all ranges in data
}
\description{
Algorithm to pick up a vector and find all ranges in it
}
\examples{
identify_ranges(c(1:3, 5, 7:8, 10:12, 14:15))
identify_ranges(c(1:3, 5, 7:8, 10:12, 14))
identify_ranges(c(1, 5, 7:8, 10:12, 14))
identify_ranges(c(1, 5, 7:8, 10:12, 14:19))
}
