% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuts_3_codes.map}
\alias{nuts_3_codes.map}
\title{NUTS_3 mapping of codes -> name}
\format{
An object of class \code{character} of length 401.
}
\usage{
nuts_3_codes.map
}
\description{
NUTS_3 mapping of codes -> name
}
\examples{
head(nuts_3_codes.map)
}
\keyword{data}
