% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{proper.cases}
\alias{proper.cases}
\title{Convrt to proper case}
\usage{
proper.cases(value, capitalize = FALSE, capitalize.all = FALSE)
}
\arguments{
\item{value}{confirmed or death strings}

\item{capitalize}{capitalize only first word}

\item{capitalize.all}{capitalize all words}
}
\value{
string with the proper mapping
}
\description{
Convrt to proper case
}
\examples{
proper.cases(c('death', 'confirmed'))
proper.cases(c('death', 'confirmed'), capitalize = TRUE)
}
