% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{range.write}
\alias{range.write}
\title{Write range for SQL query in ARCGIS}
\usage{
\method{range}{write}(a, b)
}
\arguments{
\item{a}{start of range of ObjectId}

\item{b}{end of range of ObjectId}
}
\value{
a string that can be used in a where clause of sql
}
\description{
Write range for SQL query in ARCGIS
}
\examples{
range.write(1, 10)
range.write(1, 1)
}
