% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{top30}
\alias{top30}
\title{Show top 30 cases}
\usage{
top30(dat, case.type, region.code, n = 30)
}
\arguments{
\item{dat}{data}

\item{case.type}{type of case}

\item{region.code}{state or district}

\item{n}{number to keep}
}
\value{
ggplot
}
\description{
Show top 30 cases
}
