% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{update_dataset.no.age}
\alias{update_dataset.no.age}
\title{Update dataset}
\usage{
update_dataset.no.age(force.all = FALSE)
}
\arguments{
\item{force.all}{force download of all data}
}
\value{
nothing
}
\description{
Update dataset
}
