% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{proper.cases}
\alias{proper.cases}
\title{Title}
\usage{
proper.cases(value, capitalize = FALSE, capitalize.all = FALSE)
}
\arguments{
\item{capitalize.all}{}
}
\value{

}
\description{
Title
}
\examples{
proper.cases(c('death', 'confirmed'))
proper.cases(c('death', 'confirmed'), capitalize = TRUE)
}
