% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rki.covid19}
\alias{rki.covid19}
\title{Robert Koch's Institute data for districts in Germany}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32436 rows and 13 columns.
}
\usage{
rki.covid19
}
\description{
Robert Koch's Institute data for districts in Germany
}
\examples{
head(rki.covid19)
}
\keyword{data}
